refactor: dropdown 구조 변경
- 하위 컴포넌트 분리
- provider 분리
