+++ src/component/@common/Dropdown/Dropdown.tsx
@@ -1,68 +1,14 @@
 import { ComponentProps } from 'react';
-import useOverlay from '../../../hook/useOverlay.ts';
-import {
-  dropdownItemStyle,
-  dropdownListStyle,
-  dropdownTriggerStyle,
-} from './Dropdown.styles.ts';
-import { DropdownContext, useDropdownContext } from './context.ts';
+
+import DropdownProvider from './context/dropdownProvider';
+import DropdownTrigger from './DropdownTrigger';
+import DropdownList from './DropdownList';
+import DropdownItem from './DropdownItem';
 
 interface DropdownProps extends ComponentProps<'div'> {}
 
 const Dropdown = ({ children }: DropdownProps) => {
-  const { isOpen, open, close, toggle } = useOverlay();
-
-  return (
-    <DropdownContext.Provider value={{ isOpen, open, close, toggle }}>
-      {children}
-    </DropdownContext.Provider>
-  );
-};
-
-interface DropdownTriggerProps extends ComponentProps<'div'> {}
-
-const DropdownTrigger = ({ children, ...rest }: DropdownTriggerProps) => {
-  const { toggle } = useDropdownContext();
-
-  return (
-    <div onClick={toggle} css={dropdownTriggerStyle} {...rest}>
-      {children}
-    </div>
-  );
-};
-
-interface DropdownListProps extends ComponentProps<'ul'> {}
-
-const DropdownList = ({ children, ...rest }: DropdownListProps) => {
-  const { isOpen } = useDropdownContext();
-
-  return (
-    isOpen && (
-      <ul css={dropdownListStyle} {...rest}>
-        {children}
-      </ul>
-    )
-  );
-};
-
-interface DropdownItemProps extends ComponentProps<'li'> {
-  handleClick: (content: string) => void;
-  content: string;
-}
-
-const DropdownItem = ({ content, handleClick, ...rest }: DropdownItemProps) => {
-  const { close } = useDropdownContext();
-
-  const handleSelect = () => {
-    close();
-    handleClick?.(content);
-  };
-
-  return (
-    <li onClick={handleSelect} css={dropdownItemStyle} {...rest}>
-      {content}
-    </li>
-  );
+  return <DropdownProvider>{children}</DropdownProvider>;
 };
 
 Dropdown.Root = Dropdown;
